+++ apm-agent-api/src/main/java/co/elastic/apm/api/NoopTransaction.java
@@ -101,4 +101,9 @@ enum NoopTransaction implements Transaction {
     public Span createSpan() {
         return NoopSpan.INSTANCE;
     }
+
+    @Override
+    public void injectTraceHeaders(HeaderInjector headerInjector) {
+        // noop
+    }
 }