+++ apm-agent-plugins/apm-api-plugin/src/main/java/co/elastic/apm/agent/plugin/api/ElasticApmApiInstrumentation.java
@@ -19,16 +19,16 @@
  */
 package co.elastic.apm.agent.plugin.api;
 
-import co.elastic.apm.agent.bci.ElasticApmInstrumentation;
 import co.elastic.apm.agent.bci.VisibleForAdvice;
+import co.elastic.apm.agent.impl.transaction.TraceContext;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
 import javax.annotation.Nullable;
-import java.util.Collection;
-import java.util.Collections;
+import java.lang.invoke.MethodHandle;
+import java.util.Iterator;
 
 import static net.bytebuddy.matcher.ElementMatchers.named;
 
@@ -64,6 +64,38 @@ public class ElasticApmApiInstrumentation extends ApiInstrumentation {
         private static void doStartTransaction(@Advice.Return(readOnly = false) Object transaction) {
             if (tracer != null) {
                 transaction = tracer.startTransaction();
+            }
+        }
+    }
+
+    public static class StartTransactionWithRemoteParentInstrumentation extends ElasticApmApiInstrumentation {
+
+        public StartTransactionWithRemoteParentInstrumentation() {
+            super(named("doStartTransactionWithRemoteParentFunction"));
+        }
+
+        @VisibleForAdvice
+        @Advice.OnMethodExit(suppress = Throwable.class)
+        private static void doStartTransaction(@Advice.Return(readOnly = false) Object transaction,
+                                               @Advice.Argument(0) MethodHandle getFirstHeader,
+                                               @Advice.Argument(1) @Nullable Object headerExtractor,
+                                               @Advice.Argument(2) MethodHandle getAllHeaders,
+                                               @Advice.Argument(3) @Nullable Object headersExtractor) throws Throwable {
+            if (tracer != null) {
+                if (headerExtractor != null) {
+                    final String traceparentHeader = (String) getFirstHeader.invoke(headerExtractor, TraceContext.TRACE_PARENT_HEADER);
+                    transaction = tracer.startTransaction(TraceContext.fromTraceparentHeader(), traceparentHeader);
+                } else if (headersExtractor != null) {
+                    final Iterable<String> traceparentHeader = (Iterable<String>) getAllHeaders.invoke(headersExtractor, TraceContext.TRACE_PARENT_HEADER);
+                    final Iterator<String> iterator = traceparentHeader.iterator();
+                    if (iterator.hasNext()) {
+                        transaction = tracer.startTransaction(TraceContext.fromTraceparentHeader(), iterator.next());
+                    } else {
+                        transaction = tracer.startTransaction();
+                    }
+                } else {
+                    transaction = tracer.startTransaction();
+                }
             }
         }
     }