+++ apm-agent-plugins/apm-api-plugin/src/test/java/co/elastic/apm/agent/plugin/api/SpanInstrumentationTest.java
@@ -20,12 +20,16 @@
 package co.elastic.apm.agent.plugin.api;
 
 import co.elastic.apm.agent.AbstractInstrumentationTest;
+import co.elastic.apm.agent.impl.transaction.TraceContext;
 import co.elastic.apm.api.ElasticApm;
 import co.elastic.apm.api.Scope;
 import co.elastic.apm.api.Span;
 import co.elastic.apm.api.Transaction;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -92,5 +96,36 @@ class SpanInstrumentationTest extends AbstractInstrumentationTest {
         final Transaction transaction = ElasticApm.startTransaction();
         assertThat(transaction.isSampled()).isTrue();
         assertThat(transaction.createSpan().isSampled()).isTrue();
+    }
+
+    @Test
+    void testTraceHeadersNoop() {
+        assertContainsNoTracingHeaders(ElasticApm.currentSpan());
+        assertContainsNoTracingHeaders(ElasticApm.currentTransaction());
+    }
+
+    @Test
+    void testTraceHeaders() {
+        assertContainsTracingHeaders(span);
+        assertContainsTracingHeaders(transaction);
+    }
+
+    private void assertContainsNoTracingHeaders(Span span) {
+        try (Scope scope = span.activate()) {
+            final Map<String, String> tracingHeaders = new HashMap<>();
+            span.injectTraceHeaders(tracingHeaders::put);
+            span.injectTraceHeaders(null);
+            assertThat(tracingHeaders).isEmpty();
+        }
+    }
+
+    private void assertContainsTracingHeaders(Span span) {
+        try (Scope scope = span.activate()) {
+            final Map<String, String> tracingHeaders = new HashMap<>();
+            span.injectTraceHeaders(tracingHeaders::put);
+            span.injectTraceHeaders(null);
+            final String traceparent = tracer.activeSpan().getTraceContext().getOutgoingTraceParentHeader().toString();
+            assertThat(tracingHeaders).containsEntry(TraceContext.TRACE_PARENT_HEADER, traceparent);
+        }
     }
 }